+++ Shared/Extensions/BundleExtension.swift
@@ -15,3 +15,17 @@ extension Bundle {
     self.infoDictionary?[key] as? String
   }
 }
+
+extension Bundle {
+  /// Get a string from the info plist dictionary for a given key
+  public func infoDictionaryString(forKey key: String) -> String {
+    guard let value = object(forInfoDictionaryKey: key) as? String else {
+      if AppConstants.isRunningTest {
+        /// Using dummy info dictionary while running tests, without host app set.
+        return "com.brave.tests"
+      }
+      fatalError("Could not find a string value in the Info dictionary with the key: \(key)")
+    }
+    return value
+  }
+}
